Move NiceComm to mpi_helper.py
